# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API REST Hoteles",
        "description": "API REST Hoteles",
        "contact": {
            "email": "cesarpozo25@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost/apirest-hoteles/public/",
    "basePath": "",
    "paths": {
...
        "/habitacions/{habitacion_id}/fechas": {
            "get": {
                "summary": "Get Habitacion table All Fechas",
                "parameters": [
                    {
                        "name": "Autorization",
                        "in": "header",
                        "description": "Bearer {token_access}",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "habitacion_id",
                        "in": "path",
                        "description": "un numero id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#definitions/Fecha"
                            }
                        }
                    },
                    "403": {
                        "description": "Autorization Exception",
                        "schema": {
                            "$ref": "#definitions/Errors403"
                        }
                    },
                    "404": {
                        "description": "Not Found Exception",
                        "schema": {
                            "$ref": "#definitions/Errors404"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#definitions/Errors500"
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/habitacions/{habitacion_id}/fechas/libre": {
            "get": {
                "summary": "Get Habitacion table All Fechas",
                "parameters": [
                    {
                        "name": "Autorization",
                        "in": "header",
                        "description": "Bearer {token_access}",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "habitacion_id",
                        "in": "path",
                        "description": "un numero id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "fecha_desde",
                        "in": "query",
                        "description": "una fecha de comienzo",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "fecha_hasta",
                        "in": "query",
                        "description": "una fecha de fin",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#definitions/Fecha"
                            }
                        }
                    },
                    "403": {
                        "description": "Autorization Exception",
                        "schema": {
                            "$ref": "#definitions/Errors403"
                        }
                    },
                    "404": {
                        "description": "Not Found Exception",
                        "schema": {
                            "$ref": "#definitions/Errors404"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#definitions/Errors500"
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/habitacions/{habitacion_id}/fechas/ocupado": {
            "get": {
                "summary": "Get Habitacion table All Fechas",
                "parameters": [
                    {
                        "name": "Autorization",
                        "in": "header",
                        "description": "Bearer {token_access}",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "habitacion_id",
                        "in": "path",
                        "description": "un numero id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "fecha_desde",
                        "in": "query",
                        "description": "una fecha de comienzo",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "fecha_hasta",
                        "in": "query",
                        "description": "una fecha de fin",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#definitions/Fecha"
                            }
                        }
                    },
                    "403": {
                        "description": "Autorization Exception",
                        "schema": {
                            "$ref": "#definitions/Errors403"
                        }
                    },
                    "404": {
                        "description": "Not Found Exception",
                        "schema": {
                            "$ref": "#definitions/Errors404"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#definitions/Errors500"
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
...
}